refactoring and css
