+++ src/index.css
@@ -1,0 +1,18 @@
+/* colors:
+#084F99 blue
+#E62D21 red
+#2283E6 lighter blue
+#CDE60B lime
+#89990F green
+*/
+
+.dark {
+  color: #fff;
+  background-color: #333;
+  background-color: #031c36;
+}
+.light {
+  background-color: #fff;
+  color: rgb(0, 0, 0);
+  background-color: #afd6fc;
+}
